publish notes automatically on review and sign
