+++ packages/zambdas/src/ehr/get-chart-data/helpers.ts
@@ -9,6 +9,7 @@ import {
   SearchParams,
 } from 'utils';
 import { handleCustomDTOExtractions, mapResourceToChartDataResponse } from '../../shared/chart-data';
+import { createPublishExcuseNotesOps } from '../../shared/createPublishExcuseNotesOps';
 
 type RequestOptions = ChartDataRequestedFields[keyof ChartDataRequestedFields];
 
@@ -199,6 +200,8 @@ export function convertSearchResultsToResponse(
   };
 
   const resources = parseBundleResources(bundle);
+  const documentResources = resources.filter((resource) => resource.resourceType === 'DocumentReference');
+  const publishExcuseNotesOps = createPublishExcuseNotesOps(documentResources);
 
   const chartDataResources: Resource[] = [];
   resources.forEach((resource) => {
@@ -218,5 +221,6 @@ export function convertSearchResultsToResponse(
   return {
     chartData: { ...getChartDataResponse },
     chartResources: chartDataResources,
+    publishExcuseNotesOps,
   };
 }
