+++ packages/zambdas/src/ehr/get-chart-data/index.ts
@@ -1,6 +1,6 @@
-import Oystehr, { BatchInputGetRequest, Bundle } from '@oystehr/sdk';
+import Oystehr, { BatchInputGetRequest, BatchInputRequest, Bundle } from '@oystehr/sdk';
 import { APIGatewayProxyResult } from 'aws-lambda';
-import { FhirResource, Resource } from 'fhir/r4b';
+import { DocumentReference, FhirResource, Resource } from 'fhir/r4b';
 import { ChartDataFields, ChartDataRequestedFields, GetChartDataResponse } from 'utils';
 import { checkOrCreateM2MClientToken, getPatientEncounter, ZambdaInput } from '../../shared';
 import { createOystehrClient } from '../../shared/helpers';
@@ -46,6 +46,7 @@ export async function getChartData(
 ): Promise<{
   response: GetChartDataResponse;
   chartResources: Resource[];
+  publishExcuseNotesOps: BatchInputRequest<DocumentReference>[];
 }> {
   console.time('check');
 
@@ -189,5 +190,6 @@ export async function getChartData(
   return {
     response: chartDataResult.chartData,
     chartResources: chartDataResult.chartResources,
+    publishExcuseNotesOps: chartDataResult.publishExcuseNotesOps,
   };
 }
